Moved the Zipkin library to common/tracing/zipkin. Removed precompiled header files. (#430)
