+++ source/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_subdirectory(common)
 add_subdirectory(exe)
 add_subdirectory(server)
-add_subdirectory(zipkin)
