+++ source/common/tracing/zipkin/zipkin_core_constants.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <string>
-
 namespace Zipkin {
 
 class ZipkinCoreConstants {
