+++ source/common/tracing/zipkin/zipkin_json_field_names.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <string>
-
 namespace Zipkin {
 
 class ZipkinJsonFieldNames {
