+++ test/CMakeLists.txt
@@ -26,7 +26,6 @@ include_directories(${ENVOY_LIBEVENT_INCLUDE_DIR})
 add_executable(envoy-test
   $<TARGET_OBJECTS:envoy-server>
   $<TARGET_OBJECTS:envoy-common>
-  $<TARGET_OBJECTS:zipkin>
   ${ENVOY_TEST_EXTRA_OBJECTS}
   common/access_log/access_log_manager_impl_test.cc
   common/api/api_impl_test.cc
@@ -101,6 +100,11 @@ add_executable(envoy-test
   common/stats/thread_local_store_test.cc
   common/tracing/http_tracer_impl_test.cc
   common/tracing/lightstep_tracer_impl_test.cc
+  common/tracing/zipkin/span_buffer_test.cc
+  common/tracing/zipkin/span_context_test.cc
+  common/tracing/zipkin/util_test.cc
+  common/tracing/zipkin/tracer_test.cc
+  common/tracing/zipkin/zipkin_core_types_test.cc
   common/upstream/cds_api_impl_test.cc
   common/upstream/cluster_manager_impl_test.cc
   common/upstream/health_checker_impl_test.cc
@@ -161,12 +165,7 @@ add_executable(envoy-test
   test_common/network_utility.cc
   test_common/network_utility_test.cc
   test_common/printers.cc
-  test_common/utility.cc
-  zipkin/span_buffer_test.cc
-  zipkin/span_context_test.cc
-  zipkin/util_test.cc
-  zipkin/tracer_test.cc
-  zipkin/zipkin_core_types_test.cc)
+  test_common/utility.cc)
 
 # The MOCK_METHOD* macros from gtest triggers this clang warning and it's hard
 # to work around, so we just ignore it.
