+++ test/common/tracing/zipkin/span_context_test.cc
@@ -1,5 +1,5 @@
-#include "zipkin/span_context.h"
-#include "zipkin/zipkin_core_constants.h"
+#include "common/tracing/zipkin/span_context.h"
+#include "common/tracing/zipkin/zipkin_core_constants.h"
 
 #include "gtest/gtest.h"
