+++ test/common/tracing/zipkin/tracer_test.cc
@@ -1,6 +1,6 @@
-#include "zipkin/tracer.h"
-#include "zipkin/util.h"
-#include "zipkin/zipkin_core_constants.h"
+#include "common/tracing/zipkin/tracer.h"
+#include "common/tracing/zipkin/util.h"
+#include "common/tracing/zipkin/zipkin_core_constants.h"
 
 #include "gtest/gtest.h"
